refactor(uniffi): extract fire-uniffi-messagebus
Phase 3c of the uniffi multi-namespace split.

Rust:
- Add rust/crates/fire-uniffi-messagebus with
  setup_scaffolding!("fire_uniffi_messagebus"). Hosts FireMessageBusHandle,
  the messagebus record types, and the MessageBusEventHandler callback
  interface. Eight methods move off FireAppCore verbatim: subscribe_channel,
  unsubscribe_channel, start/stop_message_bus, topic_reply_presence_state,
  bootstrap_topic_reply_presence, update_topic_reply_presence, and
  poll_notification_alert_once (which consumes NotificationAlertPollResultState
  defined in the messagebus namespace).
- FireAppCore gains messagebus: Arc<FireMessageBusHandle> + messagebus()
  getter. state_messagebus.rs is deleted.
- Promote TopicListKindState into fire-uniffi-types because MessageBusEventState
  references it to tag which topic-list rerender a message applies to.
  state_topic_list.rs keeps only TopicListQueryState (still a topics-domain
  input; will move with that domain).

iOS:
- FireSessionStore.swift routes all messagebus / presence / notification-alert
  poll calls through core.messagebus().

Build plumbing:
- native/ios-app/project.yml preBuildScript outputs gain
  fire_uniffi_messagebus.swift. Regenerated project.pbxproj.

No Android call sites reference the messagebus namespace directly today; the
existing FireSessionStore wrappers keep the rename transparent.

cargo test --workspace stays green.
