+++ native/ios-app/project.yml
@@ -109,6 +109,7 @@ targets:
         outputFiles:
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi.swift
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_diagnostics.swift
+          - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_messagebus.swift
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_notifications.swift
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_search.swift
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_types.swift
